requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format=requirements-txt --no-dev --no-hashes --no-editable --no-emit-project --output-file=requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
argcomplete==3.5.3
attrs==25.1.0
audioop-lts==0.2.1 ; python_full_version >= '3.13'
automat==24.8.1
bandit==1.8.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1 ; platform_python_implementation != 'PyPy'
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
commitizen==4.2.1
constantly==23.10.4
coverage==7.6.12
cryptography==44.0.1
cssselect==1.2.0
daily-python==0.14.2
decli==0.6.2
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
email-validator==2.2.0
execnet==2.1.1
fastapi==0.115.8
fastapi-cli==0.0.7
filelock==3.17.0
flatbuffers==25.2.10
frozenlist==1.5.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-discoveryengine==0.13.5
google-cloud-texttospeech==2.24.0
google-genai==0.7.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
humanfriendly==10.0
hyperlink==21.0.0
identify==2.6.7
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
llvmlite==0.44.0
loguru==0.7.3
lxml==5.3.1
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy==1.15.0
mypy-extensions==1.0.0
nodeenv==1.9.1
numba==0.61.0
numpy==1.26.4
onnxruntime==1.20.1
openai==1.59.9
orjson==3.10.15
packaging==24.2
parsel==1.10.0
pbr==6.1.1
pdoc==15.0.1
pillow==11.1.0
pipecat-ai==0.0.56
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==4.1.0
prompt-toolkit==3.0.50
propcache==0.2.1
protego==0.4.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22 ; platform_python_implementation != 'PyPy'
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pydispatcher==2.0.7 ; platform_python_implementation == 'CPython'
pygments==2.19.1
pyloudnorm==0.1.1
pyopenssl==25.0.0
pyparsing==3.2.1
pypydispatcher==2.1.2 ; platform_python_implementation == 'PyPy'
pyreadline3==3.5.4 ; sys_platform == 'win32'
pytest==8.3.4
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-xdist==3.6.1
python-deepcompare==2.1.0
python-dotenv==1.0.1
python-multipart==0.0.20
pyyaml==6.0.2
questionary==2.1.0
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
resampy==0.4.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
ruff==0.9.6
scipy==1.15.1
scrapy==2.12.0
service-identity==24.2.0
setuptools==75.8.0
shellingham==1.5.4
sniffio==1.3.1
soxr==0.5.0.post1
starlette==0.45.3
stevedore==5.4.0
sympy==1.13.3
termcolor==2.5.0
tldextract==5.1.3
tomlkit==0.13.2
tqdm==4.67.1
twisted==24.11.0
typer==0.15.1
typing-extensions==4.12.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'
virtualenv==20.29.2
w3lib==2.3.1
watchfiles==1.0.4
wcwidth==0.2.13
websockets==13.1
win32-setctime==1.2.0 ; sys_platform == 'win32'
yarl==1.18.3
zope-interface==7.2
